CSHARP-3341: Code review.
